base para los cambios generales de un nuevo comando
