+++ funcionalidad/cumpleañero.js
@@ -1,0 +1,15 @@
+class Cumpleañero {
+   // Propiedades.
+   // Constructor.
+   // Metodos.
+   definirDiasRestantes(fecha){
+      let diasRestantes = 0
+      console.log("[cumpleañero] Fecha nacimiento: ", fecha);
+      return diasRestantes;
+   }
+
+   // Getters & Setters
+   // Destructor.
+}
+
+module.exports = Cumpleañero;
